requirements.txt
fastapi==0.97.0
fasteners==0.14.1
fonttools==4.40.0
future==0.18.2
google-cloud-pubsub==2.17.1
h11==0.14.0
httplib2==0.20.2
isort==5.12.0
imbalanced-learn==0.10.1
imblearn==0.0
joblib==1.2.0
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazy-object-proxy==1.9.0
linecache2==1.0.0
lockfile==0.12.2
logilab-common==1.8.2
macaroonbakery==1.3.1
matplotlib==3.7.1
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==0.4.3
nltk==3.8.1
numpy==1.25.0
oauthlib==3.2.0
olefile==0.46
packaging==23.1
pandas==2.0.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.8.0
ptyprocess==0.7.0
pydantic==1.10.9
PyJWT==2.3.0
pylint==2.17.4
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2023.6.3
reportlab==3.6.8
requests==2.25.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.1
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
unittest2==1.1.0
update-checker==0.18.0
urllib3==1.26.5
uvicorn==0.22.0
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.19.0
websocket-client==1.6.1
websockets==11.0.3
wrapt==1.15.0
wsproto==1.0.0
xdg==5
